+++ packages/form/src/index.ts
@@ -1,6 +1,8 @@
 export * from "./Form";
 export * from "./Fieldset";
 export * from "./FormThemeProvider";
+export * from "./FormMessage";
+export * from "./FormMessageContainer";
 
 export * from "./file-input";
 export * from "./label";
