+++ packages/form/src/text-field/__tests__/TextAreaWithMessage.tsx
@@ -7,7 +7,7 @@ import { ErrorOutlineFontIcon } from "@react-md/material-icons";
 
 import { TextAreaWithMessage } from "../TextAreaWithMessage";
 import { TextFieldHookOptions, useTextField } from "../useTextField";
-import { FormMessageProps } from "../FormMessage";
+import { FormMessageProps } from "../../FormMessage";
 
 function Test({
   id = "field-id",
